allow more characters in usernames
DRF’s default lookup regex for url parameters matches everything except
slashes and dots, but especially the latter is used particularly often
in usernames. We should declare an explicit pattern for the username
in order to avoid url-reverse errors for usernames with dots, but also
restrict the range of characters that are allowed to use.